setgroups: new mode 'shadow' for /proc/PID/setgroups
add a new mode 'shadow' to the /proc/PID/setgroups file.

When the 'shadow' string is written to the file, the current process
groups are inherited from the process and stored into the user
namespace.  These groups will be silently added on each setgroups(2)
call.  A child user namespace won't be able to drop these groups
anymore.

It enables using setgroups(2) in user namespaces on systems where
negative ACLs are used.

Once the 'shadow' bit is set, a child user namespace cannot drop it
anymore.  A child user namespace will inherit the list of shadowed
groups at creation.

A child user namespace can reset the list of shadow groups by writing
again the string 'shadow' to the setgroups file.  It will add all the
additional groups that are not yet into the shadow groups set.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>