overlay: add nosync mount option
By default, overlay skips sync on files residing on a lower layer.
Add a new mount option "nosync" to skip sync operations also for files
residing on the upper layer.

It is very useful for running containers that are removed on
termination so it is pointless to sync files.

On my machine "yum install -y emacs" passes from:
        User time (seconds): 0.15
        System time (seconds): 0.38
        Percent of CPU this job got: 0%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 3:56.35
to:
        User time (seconds): 0.19
        System time (seconds): 0.39
        Percent of CPU this job got: 0%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 1:09.60

and "apt install -y emacs" passes from:
        User time (seconds): 0.28
        System time (seconds): 0.57
        Percent of CPU this job got: 0%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 3:57.65
to:
        User time (seconds): 0.29
        System time (seconds): 0.62
        Percent of CPU this job got: 0%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 1:51.86

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>