kernel: PR_RESTRICT_SELF_EXE blocks PR_SET_DUMPABLE
Once prctl(PR_RESTRICT_SELF_EXE) is set, it will blocks
PR_SET_DUMPABLE to be set again.  It is not done automatically so that
PR_SET_DUMPABLE can be set at a later time if needed, e.g. a container
runtime might want to do that after the user namespace is joined.

It is currently kind of expected that a container uses execve to pass
control to a binary in the container chroot, since the /proc/self/exe
and /proc/self/map_files refer to files that are not part of the
container itself; so if the process can be compromised in any way
(e.g. an interpreter running a script in the container without calling
execve), it could potentially lead to host compromise.  It is
necessary to ensure that the current process cannot be set to dumpable
anymore, since that would allow another process in the container to
get access to its proc files.

A container runtime could use the new prctl in the following way:

- run prctl(PR_RESTRICT_SELF_EXE) as soon as possible.
- if prctl(PR_RESTRICT_SELF_EXE) fails then use the more expensive current
  fallbacks mechanism.
- perform all operations that require the process to be dumpable.
- prctl(PR_SET_DUMPABLE, 0), as it is already done today.
- execve the new binary or run directly the interpreter.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>